requirements.txt
annotated-types==0.7.0
anyio==4.12.1
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.4
cssselect2==0.8.0
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
extract-video-ppt==1.1.2
filelock==3.20.3
fonttools==4.61.1
fpdf2==2.8.4
fsspec==2026.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
kiwisolver==1.4.9
llvmlite==0.46.0
Markdown==3.10.1
MarkupSafe==3.0.3
matplotlib==3.10.8
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.17.0
openai-whisper==20250625
opencv-python==4.13.0.92
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
pyparsing==3.3.2
pypdfium2==5.3.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
regex==2026.1.15
requests==2.32.5
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
torch==2.10.0
tqdm==4.67.3
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
weasyprint==68.0
webencodings==0.5.1
wheel==0.46.3
zopfli==0.4.0
